Refactor backup and restore operations into graph operations module
This commit removes the existing `create_backup.py` and `restore_backup.py` files and introduces new modules under `graph_operations` for handling backup and restore operations. The new structure enhances organization and aligns with the overall architecture of the API.

Key Changes:
- Deleted `create_backup.py` and `restore_backup.py`.
- Added `op_create_backup.py` and `op_restore_backup.py` to `graph_operations` for improved modularity.
- Updated response handling and request structures to support new operation envelopes and error models.

This refactor aims to streamline the backup and restore processes, making them more consistent with other graph operations.
