--- robosystems_client/api/subgraphs/create_subgraph.py
@@ -1,292 +1,0 @@
-from http import HTTPStatus
-from typing import Any
-from urllib.parse import quote
-
-import httpx
-
-from ... import errors
-from ...client import AuthenticatedClient, Client
-from ...models.create_subgraph_request import CreateSubgraphRequest
-from ...models.http_validation_error import HTTPValidationError
-from ...types import Response
-
-
-def _get_kwargs(
-  graph_id: str,
-  *,
-  body: CreateSubgraphRequest,
-) -> dict[str, Any]:
-  headers: dict[str, Any] = {}
-
-  _kwargs: dict[str, Any] = {
-    "method": "post",
-    "url": "/v1/graphs/{graph_id}/subgraphs".format(
-      graph_id=quote(str(graph_id), safe=""),
-    ),
-  }
-
-  _kwargs["json"] = body.to_dict()
-
-  headers["Content-Type"] = "application/json"
-
-  _kwargs["headers"] = headers
-  return _kwargs
-
-
-def _parse_response(
-  *, client: AuthenticatedClient | Client, response: httpx.Response
-) -> Any | HTTPValidationError | None:
-  if response.status_code == 200:
-    response_200 = response.json()
-    return response_200
-
-  if response.status_code == 422:
-    response_422 = HTTPValidationError.from_dict(response.json())
-
-    return response_422
-
-  if client.raise_on_unexpected_status:
-    raise errors.UnexpectedStatus(response.status_code, response.content)
-  else:
-    return None
-
-
-def _build_response(
-  *, client: AuthenticatedClient | Client, response: httpx.Response
-) -> Response[Any | HTTPValidationError]:
-  return Response(
-    status_code=HTTPStatus(response.status_code),
-    content=response.content,
-    headers=response.headers,
-    parsed=_parse_response(client=client, response=response),
-  )
-
-
-def sync_detailed(
-  graph_id: str,
-  *,
-  client: AuthenticatedClient,
-  body: CreateSubgraphRequest,
-) -> Response[Any | HTTPValidationError]:
-  """Create Subgraph
-
-   Create a new subgraph within a parent graph, with optional data forking.
-
-  **Requirements:**
-  - Valid authentication
-  - Parent graph must exist and be accessible to the user
-  - User must have 'admin' permission on the parent graph
-  - Parent graph tier must support subgraphs (LadybugDB Large/XLarge)
-  - Must be within subgraph quota limits
-  - Subgraph name must be unique within the parent graph
-
-  **Fork Mode:**
-  When `fork_parent=true`, the operation:
-  - Returns immediately with an operation_id for SSE monitoring
-  - Copies data from parent graph to the new subgraph
-  - Supports selective forking via metadata.fork_options
-  - Tracks progress in real-time via SSE
-
-  **Returns:**
-  - Without fork: Immediate SubgraphResponse with created subgraph details
-  - With fork: Operation response with SSE monitoring endpoint
-
-  **Subgraph ID format:** `{parent_id}_{subgraph_name}` (e.g., kg1234567890abcdef_dev)
-
-  **Usage:**
-  - Subgraphs share parent's credit pool
-  - Subgraph ID can be used in all standard `/v1/graphs/{graph_id}/*` endpoints
-  - Permissions inherited from parent graph
-
-  Args:
-      graph_id (str):
-      body (CreateSubgraphRequest): Request model for creating a subgraph.
-
-  Raises:
-      errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
-      httpx.TimeoutException: If the request takes longer than Client.timeout.
-
-  Returns:
-      Response[Any | HTTPValidationError]
-  """
-
-  kwargs = _get_kwargs(
-    graph_id=graph_id,
-    body=body,
-  )
-
-  response = client.get_httpx_client().request(
-    **kwargs,
-  )
-
-  return _build_response(client=client, response=response)
-
-
-def sync(
-  graph_id: str,
-  *,
-  client: AuthenticatedClient,
-  body: CreateSubgraphRequest,
-) -> Any | HTTPValidationError | None:
-  """Create Subgraph
-
-   Create a new subgraph within a parent graph, with optional data forking.
-
-  **Requirements:**
-  - Valid authentication
-  - Parent graph must exist and be accessible to the user
-  - User must have 'admin' permission on the parent graph
-  - Parent graph tier must support subgraphs (LadybugDB Large/XLarge)
-  - Must be within subgraph quota limits
-  - Subgraph name must be unique within the parent graph
-
-  **Fork Mode:**
-  When `fork_parent=true`, the operation:
-  - Returns immediately with an operation_id for SSE monitoring
-  - Copies data from parent graph to the new subgraph
-  - Supports selective forking via metadata.fork_options
-  - Tracks progress in real-time via SSE
-
-  **Returns:**
-  - Without fork: Immediate SubgraphResponse with created subgraph details
-  - With fork: Operation response with SSE monitoring endpoint
-
-  **Subgraph ID format:** `{parent_id}_{subgraph_name}` (e.g., kg1234567890abcdef_dev)
-
-  **Usage:**
-  - Subgraphs share parent's credit pool
-  - Subgraph ID can be used in all standard `/v1/graphs/{graph_id}/*` endpoints
-  - Permissions inherited from parent graph
-
-  Args:
-      graph_id (str):
-      body (CreateSubgraphRequest): Request model for creating a subgraph.
-
-  Raises:
-      errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
-      httpx.TimeoutException: If the request takes longer than Client.timeout.
-
-  Returns:
-      Any | HTTPValidationError
-  """
-
-  return sync_detailed(
-    graph_id=graph_id,
-    client=client,
-    body=body,
-  ).parsed
-
-
-async def asyncio_detailed(
-  graph_id: str,
-  *,
-  client: AuthenticatedClient,
-  body: CreateSubgraphRequest,
-) -> Response[Any | HTTPValidationError]:
-  """Create Subgraph
-
-   Create a new subgraph within a parent graph, with optional data forking.
-
-  **Requirements:**
-  - Valid authentication
-  - Parent graph must exist and be accessible to the user
-  - User must have 'admin' permission on the parent graph
-  - Parent graph tier must support subgraphs (LadybugDB Large/XLarge)
-  - Must be within subgraph quota limits
-  - Subgraph name must be unique within the parent graph
-
-  **Fork Mode:**
-  When `fork_parent=true`, the operation:
-  - Returns immediately with an operation_id for SSE monitoring
-  - Copies data from parent graph to the new subgraph
-  - Supports selective forking via metadata.fork_options
-  - Tracks progress in real-time via SSE
-
-  **Returns:**
-  - Without fork: Immediate SubgraphResponse with created subgraph details
-  - With fork: Operation response with SSE monitoring endpoint
-
-  **Subgraph ID format:** `{parent_id}_{subgraph_name}` (e.g., kg1234567890abcdef_dev)
-
-  **Usage:**
-  - Subgraphs share parent's credit pool
-  - Subgraph ID can be used in all standard `/v1/graphs/{graph_id}/*` endpoints
-  - Permissions inherited from parent graph
-
-  Args:
-      graph_id (str):
-      body (CreateSubgraphRequest): Request model for creating a subgraph.
-
-  Raises:
-      errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
-      httpx.TimeoutException: If the request takes longer than Client.timeout.
-
-  Returns:
-      Response[Any | HTTPValidationError]
-  """
-
-  kwargs = _get_kwargs(
-    graph_id=graph_id,
-    body=body,
-  )
-
-  response = await client.get_async_httpx_client().request(**kwargs)
-
-  return _build_response(client=client, response=response)
-
-
-async def asyncio(
-  graph_id: str,
-  *,
-  client: AuthenticatedClient,
-  body: CreateSubgraphRequest,
-) -> Any | HTTPValidationError | None:
-  """Create Subgraph
-
-   Create a new subgraph within a parent graph, with optional data forking.
-
-  **Requirements:**
-  - Valid authentication
-  - Parent graph must exist and be accessible to the user
-  - User must have 'admin' permission on the parent graph
-  - Parent graph tier must support subgraphs (LadybugDB Large/XLarge)
-  - Must be within subgraph quota limits
-  - Subgraph name must be unique within the parent graph
-
-  **Fork Mode:**
-  When `fork_parent=true`, the operation:
-  - Returns immediately with an operation_id for SSE monitoring
-  - Copies data from parent graph to the new subgraph
-  - Supports selective forking via metadata.fork_options
-  - Tracks progress in real-time via SSE
-
-  **Returns:**
-  - Without fork: Immediate SubgraphResponse with created subgraph details
-  - With fork: Operation response with SSE monitoring endpoint
-
-  **Subgraph ID format:** `{parent_id}_{subgraph_name}` (e.g., kg1234567890abcdef_dev)
-
-  **Usage:**
-  - Subgraphs share parent's credit pool
-  - Subgraph ID can be used in all standard `/v1/graphs/{graph_id}/*` endpoints
-  - Permissions inherited from parent graph
-
-  Args:
-      graph_id (str):
-      body (CreateSubgraphRequest): Request model for creating a subgraph.
-
-  Raises:
-      errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
-      httpx.TimeoutException: If the request takes longer than Client.timeout.
-
-  Returns:
-      Any | HTTPValidationError
-  """
-
-  return (
-    await asyncio_detailed(
-      graph_id=graph_id,
-      client=client,
-      body=body,
-    )
-  ).parsed
