+++ robosystems_client/models/operation_envelope.py
@@ -18,40 +18,33 @@ T = TypeVar("T", bound="OperationEnvelope")
 
 @_attrs_define
 class OperationEnvelope:
-  """Uniform response shape for every extensions operation endpoint.
+  """Uniform response shape for every operation endpoint.
 
-  Every dispatch through `/extensions/{domain}/{graph_id}/operations/{op}`
-  returns an envelope carrying an `op_<ULID>` operation_id. That id is the
-  bridge to the platform's monitoring surface: pass it to
-  `GET /v1/operations/{operation_id}/stream` (see `routers/operations.py`)
-  to subscribe to SSE progress events. Sync commands complete in the
-  envelope itself; async commands (`status: "pending"`, HTTP 202) hand off
-  to a background worker and stream their tail through the same SSE
-  endpoint until completion. Failed dispatches still mint an `operation_id`
-  so the audit log and any partial SSE events stay correlatable.
+  Every dispatch through an operation surface returns an envelope carrying
+  an ``op_<ULID>`` operation_id.  That id is the bridge to the platform's
+  monitoring surface: pass it to
+  ``GET /v1/operations/{operation_id}/stream`` (see ``routers/operations.py``)
+  to subscribe to SSE progress events.  Sync commands complete in the
+  envelope itself; async commands (``status: "pending"``, HTTP 202) hand
+  off to a background worker and stream their tail through the same SSE
+  endpoint until completion.  Failed dispatches still mint an
+  ``operation_id`` so the audit log and any partial SSE events stay
+  correlatable.
 
   Fields:
-  - `operation`: kebab-case command name (e.g. `close-period`)
-  - `operation_id`: `op_`-prefixed ULID; always present, usable for audit
-    correlation and — for async commands — SSE subscription via
-    `/v1/operations/{operation_id}/stream`
-  - `status`: `"completed"` (sync, HTTP 200), `"pending"` (async, HTTP 202),
-    or `"failed"` (error responses)
-  - `result`: the domain-specific payload (the original Pydantic response)
-    or `None` for async/failed cases
-  - `at`: ISO-8601 UTC timestamp of when the envelope was minted (for sync
-    ops this is the completion time; for async/pending it's the enqueue time)
-  - `created_by`: user ID of the caller who initiated this operation, for
-    audit correlation without having to cross-reference the audit log.
-    Always populated for dispatcher-routed calls; may be `None` for legacy
-    direct `wrap_completed(...)` callers.
-  - `idempotent_replay`: `True` when the dispatcher returned this envelope
-    from the idempotency cache (the underlying command did NOT execute
-    again). `False` on every fresh execution. Clients can use this to
-    distinguish "my retry succeeded" from "the server re-ran the command"
-    without having to track their own request identity. The metrics
-    decorator also reads this attribute to suppress business-event counter
-    increments on replays so dashboards stay honest.
+  - ``operation``: kebab-case command name (e.g. ``close-period``)
+  - ``operation_id``: ``op_``-prefixed ULID; always present, usable for
+    audit correlation and — for async commands — SSE subscription via
+    ``/v1/operations/{operation_id}/stream``
+  - ``status``: ``"completed"`` (sync, HTTP 200), ``"pending"``
+    (async, HTTP 202), or ``"failed"`` (error responses)
+  - ``result``: the domain-specific payload (the original Pydantic
+    response) or ``None`` for async/failed cases
+  - ``at``: ISO-8601 UTC timestamp of when the envelope was minted
+  - ``created_by``: user ID of the caller who initiated this operation
+  - ``idempotent_replay``: ``True`` when the dispatcher returned this
+    envelope from the idempotency cache (the underlying command did NOT
+    execute again)
 
       Attributes:
           operation (str): Kebab-case operation name
