+++ tests/test_materialization_client.py
@@ -1,12 +1,11 @@
-"""Unit tests for MaterializationClient."""
+"""Unit tests for GraphClient materialization."""
 
 import pytest
 from unittest.mock import Mock, patch
-from robosystems_client.clients.materialization_client import (
-  MaterializationClient,
+from robosystems_client.clients.graph_client import (
+  GraphClient,
   MaterializationOptions,
   MaterializationResult,
-  MaterializationStatus,
 )
 from robosystems_client.clients.operation_client import (
   OperationResult,
@@ -86,32 +85,14 @@ class TestMaterializationDataclasses:
     assert result.success is False
     assert result.error == "Connection timeout"
 
-  def test_materialization_status(self):
-    """Test MaterializationStatus dataclass."""
-    status = MaterializationStatus(
-      graph_id="graph-123",
-      is_stale=True,
-      stale_reason="Files uploaded since last materialization",
-      stale_since="2025-01-15T10:00:00Z",
-      last_materialized_at="2025-01-14T08:00:00Z",
-      materialization_count=5,
-      hours_since_materialization=26.0,
-      message="Graph is stale",
-    )
-
-    assert status.graph_id == "graph-123"
-    assert status.is_stale is True
-    assert status.materialization_count == 5
-    assert status.hours_since_materialization == 26.0
-
 
 @pytest.mark.unit
-class TestMaterializationClientInit:
-  """Test suite for MaterializationClient initialization."""
+class TestGraphClientMaterializeInit:
+  """Test suite for GraphClient initialization relevant to materialization."""
 
   def test_client_initialization(self, mock_config):
     """Test that client initializes correctly with config."""
-    client = MaterializationClient(mock_config)
+    client = GraphClient(mock_config)
 
     assert client.base_url == "http://localhost:8000"
     assert client.token == "test-api-key"
@@ -120,7 +101,7 @@ class TestMaterializationClientInit:
 
   def test_operation_client_lazy_creation(self, mock_config):
     """Test that operation client is created lazily."""
-    client = MaterializationClient(mock_config)
+    client = GraphClient(mock_config)
 
     assert client._operation_client is None
     op_client = client.operation_client
@@ -131,18 +112,19 @@ class TestMaterializationClientInit:
 
 @pytest.mark.unit
 class TestMaterialize:
-  """Test suite for MaterializationClient.materialize method."""
+  """Test suite for GraphClient.materialize method."""
 
-  @patch("robosystems_client.clients.materialization_client.materialize_graph")
-  def test_materialize_success(self, mock_mat, mock_config, graph_id):
+  @patch("robosystems_client.api.graph_operations.op_materialize.sync_detailed")
+  @patch.object(GraphClient, "_get_authenticated_client")
+  def test_materialize_success(self, mock_auth, mock_mat, mock_config, graph_id):
     """Test successful materialization."""
-    # Mock initial response with operation_id
+    mock_auth.return_value = Mock()
+
     mock_resp = Mock()
-    mock_resp.status_code = 200
+    mock_resp.status_code = 202
     mock_resp.parsed = Mock(operation_id="op-mat-123")
     mock_mat.return_value = mock_resp
 
-    # Mock the operation client monitoring
     op_result = OperationResult(
       operation_id="op-mat-123",
       status=OperationStatus.COMPLETED,
@@ -157,7 +139,7 @@ class TestMaterialize:
       execution_time_ms=2000.0,
     )
 
-    client = MaterializationClient(mock_config)
+    client = GraphClient(mock_config)
     client._operation_client = Mock()
     client._operation_client.monitor_operation.return_value = op_result
 
@@ -168,26 +150,34 @@ class TestMaterialize:
     assert result.tables_materialized == ["Entity"]
     assert result.total_rows == 100
 
-  @patch("robosystems_client.clients.materialization_client.materialize_graph")
-  def test_materialize_api_failure(self, mock_mat, mock_config, graph_id):
+  @patch("robosystems_client.api.graph_operations.op_materialize.sync_detailed")
+  @patch.object(GraphClient, "_get_authenticated_client")
+  def test_materialize_api_failure(self, mock_auth, mock_mat, mock_config, graph_id):
     """Test materialization when API returns error."""
+    mock_auth.return_value = Mock()
+
     mock_resp = Mock()
     mock_resp.status_code = 500
     mock_resp.parsed = None
     mock_resp.content = b'{"detail": "Internal error"}'
     mock_mat.return_value = mock_resp
 
-    client = MaterializationClient(mock_config)
+    client = GraphClient(mock_config)
     result = client.materialize(graph_id)
 
     assert result.success is False
     assert result.status == "failed"
 
-  @patch("robosystems_client.clients.materialization_client.materialize_graph")
-  def test_materialize_operation_failed(self, mock_mat, mock_config, graph_id):
+  @patch("robosystems_client.api.graph_operations.op_materialize.sync_detailed")
+  @patch.object(GraphClient, "_get_authenticated_client")
+  def test_materialize_operation_failed(
+    self, mock_auth, mock_mat, mock_config, graph_id
+  ):
     """Test materialization when operation fails."""
+    mock_auth.return_value = Mock()
+
     mock_resp = Mock()
-    mock_resp.status_code = 200
+    mock_resp.status_code = 202
     mock_resp.parsed = Mock(operation_id="op-fail")
     mock_mat.return_value = mock_resp
 
@@ -198,7 +188,7 @@ class TestMaterialize:
       execution_time_ms=1000.0,
     )
 
-    client = MaterializationClient(mock_config)
+    client = GraphClient(mock_config)
     client._operation_client = Mock()
     client._operation_client.monitor_operation.return_value = op_result
 
@@ -210,18 +200,21 @@ class TestMaterialize:
   def test_materialize_no_token(self, mock_config, graph_id):
     """Test materialize fails without API key."""
     mock_config["token"] = None
-    client = MaterializationClient(mock_config)
+    client = GraphClient(mock_config)
 
     result = client.materialize(graph_id)
 
     assert result.success is False
     assert "No API key" in result.error
 
-  @patch("robosystems_client.clients.materialization_client.materialize_graph")
-  def test_materialize_with_progress(self, mock_mat, mock_config, graph_id):
+  @patch("robosystems_client.api.graph_operations.op_materialize.sync_detailed")
+  @patch.object(GraphClient, "_get_authenticated_client")
+  def test_materialize_with_progress(self, mock_auth, mock_mat, mock_config, graph_id):
     """Test materialize calls progress callback."""
+    mock_auth.return_value = Mock()
+
     mock_resp = Mock()
-    mock_resp.status_code = 200
+    mock_resp.status_code = 202
     mock_resp.parsed = Mock(operation_id="op-progress")
     mock_mat.return_value = mock_resp
 
@@ -243,7 +236,7 @@ class TestMaterialize:
       on_progress=lambda msg: progress_messages.append(msg)
     )
 
-    client = MaterializationClient(mock_config)
+    client = GraphClient(mock_config)
     client._operation_client = Mock()
     client._operation_client.monitor_operation.return_value = op_result
 
@@ -251,11 +244,14 @@ class TestMaterialize:
 
     assert len(progress_messages) >= 2  # At least "Submitting" and "queued"
 
-  @patch("robosystems_client.clients.materialization_client.materialize_graph")
-  def test_materialize_with_rebuild(self, mock_mat, mock_config, graph_id):
+  @patch("robosystems_client.api.graph_operations.op_materialize.sync_detailed")
+  @patch.object(GraphClient, "_get_authenticated_client")
+  def test_materialize_with_rebuild(self, mock_auth, mock_mat, mock_config, graph_id):
     """Test materialize passes rebuild option."""
+    mock_auth.return_value = Mock()
+
     mock_resp = Mock()
-    mock_resp.status_code = 200
+    mock_resp.status_code = 202
     mock_resp.parsed = Mock(operation_id="op-rebuild")
     mock_mat.return_value = mock_resp
 
@@ -272,7 +268,7 @@ class TestMaterialize:
       execution_time_ms=5000.0,
     )
 
-    client = MaterializationClient(mock_config)
+    client = GraphClient(mock_config)
     client._operation_client = Mock()
     client._operation_client.monitor_operation.return_value = op_result
 
@@ -280,57 +276,7 @@ class TestMaterialize:
     result = client.materialize(graph_id, options)
 
     assert result.success is True
-    # Verify the request body had rebuild=True
+    # Verify the call had rebuild=True and force=True
     call_kwargs = mock_mat.call_args[1]
-    assert call_kwargs["body"].rebuild is True
-    assert call_kwargs["body"].force is True
-
-
-@pytest.mark.unit
-class TestMaterializationStatus:
-  """Test suite for MaterializationClient.status method."""
-
-  @patch("robosystems_client.clients.materialization_client.get_materialization_status")
-  def test_get_status(self, mock_status, mock_config, graph_id):
-    """Test getting materialization status."""
-    mock_resp = Mock()
-    mock_resp.status_code = 200
-    mock_resp.parsed = Mock(
-      graph_id=graph_id,
-      is_stale=True,
-      stale_reason="New files uploaded",
-      stale_since="2025-01-15T10:00:00Z",
-      last_materialized_at="2025-01-14T08:00:00Z",
-      materialization_count=3,
-      hours_since_materialization=26.0,
-      message="Graph is stale",
-    )
-    mock_status.return_value = mock_resp
-
-    client = MaterializationClient(mock_config)
-    status = client.status(graph_id)
-
-    assert status is not None
-    assert status.is_stale is True
-    assert status.materialization_count == 3
-
-  @patch("robosystems_client.clients.materialization_client.get_materialization_status")
-  def test_get_status_failure(self, mock_status, mock_config, graph_id):
-    """Test status returns None on failure."""
-    mock_resp = Mock()
-    mock_resp.status_code = 500
-    mock_resp.parsed = None
-    mock_status.return_value = mock_resp
-
-    client = MaterializationClient(mock_config)
-    status = client.status(graph_id)
-
-    assert status is None
-
-  def test_status_no_token(self, mock_config, graph_id):
-    """Test status returns None without token."""
-    mock_config["token"] = None
-    client = MaterializationClient(mock_config)
-    status = client.status(graph_id)
-
-    assert status is None
+    assert call_kwargs["rebuild"] is True
+    assert call_kwargs["force"] is True
